file syncing

+++ app/clide/collaboration/Server.scala
@@ -15,11 +15,11 @@ class Server(initialState: Document) {
   def revision = history.length
   def getHistory = history.view
     
-  def applyOperation(rev: Int, operation: Operation): Try[Operation] = {
+  def applyOperation(rev: Long, operation: Operation): Try[Operation] = {
     val result = for {
 	  concurrentOps <- Try {
 	    require((0 to revision) contains rev, "invalid revision: " + rev)	    
-	    history.view(rev, revision) 
+	    history.view(rev.toInt, revision) // TODO: Long Revisions 
 	  }
 	  operation <- concurrentOps.foldLeft(Success(operation): Try[Operation]) {
 	    case (a,b) => a.flatMap(a => Operation.transform(a,b).map(_._1)) }
